Transaction fully functional. Changes in java files from int to double based on DTO changes.

+++ ira_bank/src/edu/irabank/controller/TransactionController.java
@@ -61,7 +61,7 @@ import edu.irabank.service.TransactionService;
 			String Accountno = accountFormBean.getAccountNumber();
 			System.out.println("balance from formbean account" + accountFormBean.getAccountNumber());
 			System.out.println("balance from formbean" + accountFormBean.getAmount());
-			Integer balAnce = Integer.parseInt(accountFormBean.getAmount());
+			Double balAnce = Double.parseDouble(accountFormBean.getAmount());
 			
 			String type = accountFormBean.getCreditDebit();
 			boolean isAccountExist = transactionService.getAccountNumber(Accountnum);
